fixing all

+++ client/resources/script/main.js
@@ -1,32 +1,75 @@
-$("#logout").hide()
-$('.register-form').hide()
-$('.container').hide()
+$(document).ready(function () {
+    console.log('document ready!')
+    if (localStorage.token === undefined) {
+        $("#logout").hide()
+        $('.register-form').hide()
+        $('.container').hide()
+        $('.form-add-todo').hide()
+        $('#add-todo-button').hide()
+    } else {
+        hasLogin(localStorage.username)
+    }
+
+    $("#add-todo-button").click(function () {
+        $(".form-add-todo").modal("show");
+        // $("#add-todo-button").css("display", "none");
+    });
+
+    $("#add-todo-button-cancel").click(function () {
+        $(".form-add-todo").modal("hide");
+        // $("#add-todo-button").css("display", "block");
+    });
+
+    $("#form-add-todo").submit(addTodo);
+
+})
+
+function hasLogin() {
+    $('.login-form').hide()
+    $('.g-signin2').hide()
+    $('.register-form').hide()
+    $('#logout').show()
+    $('.container').show()
+    $('#add-todo-button').show()
+    $('.form-add-todo').hide()
+    $("#author").text(`Author: ${localStorage.username}`);
+    showTodos()
+}
 
 function login() {
     event.preventDefault();
-    const data = {
-        username: $('#username-login').val(),
-        password: $('#password-login').val()
-    };
+    if (!$('#username-login').val() || !$('#password-login').val()) {
+        swal({
+            title: "Please Fill All Column!",
+            icon: "error",
+            buttons: false,
+            timer: 2000
+        });
+    } else {
+        const data = {
+            username: $('#username-login').val(),
+            password: $('#password-login').val()
+        };
 
-    $.ajax({
-        url: `http://localhost:3000/api/login`,
-        type: 'POST',
-        data
-
-    })
-        .done(function (token) {
-            console.log(token)
-            localStorage.setItem('token',token)
-            $('.login-form').hide()
-            $('.g-signin2').hide()
-            $('.register-form').hide()
-            $('#logout').show()
-            $('.container').show()
+        $.ajax({
+            url: `http://localhost:3000/api/login`,
+            type: 'POST',
+            data
         })
-        .fail(function (gg, textStatus) {
-            console.log('Err:', textStatus)
-        })
+            .done(function ({ token, username }) {
+                localStorage.setItem('token', token)
+                localStorage.setItem('username', username)
+                swal("Login Success!", `Welcome back, ${username}!`, "success")
+                hasLogin()
+            })
+            .fail(function (gg, textStatus) {
+                console.log('Err:', textStatus)
+                swal({
+                    title: "User doesn't exist. Please register first!",
+                    icon: "error",
+                });
+            })
+    }
 }
 
 function onSignIn(googleUser) {
@@ -39,13 +82,11 @@ function onSignIn(googleUser) {
         }
     })
         .done(function (data) {
-            console.log('Sign in as', data);
-            localStorage.setItem('token',data)
-            $('.login-form').hide()
-            $('.g-signin2').hide()
-            $('.register-form').hide()
-            $('#logout').show()
-            $('.container').show()
+            const { token, username } = data
+            localStorage.setItem('token', token)
+            localStorage.setItem('username', username)
+            swal("Good Job!", `Welcome to the todos app, ${username}!`, "success");
+            hasLogin()
         })
         .fail(function (gg, textStatus) {
             console.log('Error', textStatus);
@@ -53,19 +94,43 @@ function onSignIn(googleUser) {
 }
 
 function signOut() {
-    localStorage.clear()
-    var auth2 = gapi.auth2.getAuthInstance();
-    auth2.signOut().then(function () {
-        console.log('User signed out.');
-    });
-    $("#logout").hide()
-    $('.register-form').hide()
-    $('.login-form').show()
-    $('.g-signin2').show()
-    $('.container').hide()
+
+    swal({
+        title: 'Are you sure want to logout?',
+        icon: 'warning',
+        buttons: true,
+        
+    })
+        .then((result) => {
+            if (result) {
+                swal(
+                    'Good bye!',
+                    'You logged out from application!',
+                    'success'
+                )
+
+                localStorage.clear()
+                sessionStorage.clear()
+                var auth2 = gapi.auth2.getAuthInstance();
+                auth2.signOut().then(function () {
+                    console.log('User signed out.');
+                });
+                $('#username-login').val("")
+                $('#password-login').val("")
+                $('#username-register').val("")
+                $('#password-register').val("")
+                $("#logout").hide()
+                $('.register-form').hide()
+                $('.login-form').show()
+                $('.g-signin2').show()
+                $('.container').hide()
+                $('.form-add-todo').hide()
+                $('#add-todo-button').hide()
+            }
+        })
 }
 
-function showRegister(){
+function showRegister() {
     $('.register-form').show()
     $('.login-form').hide()
     $('.g-signin2').hide()
@@ -75,37 +140,47 @@ function showRegister(){
 
 function register() {
     event.preventDefault();
-    const data = {
-        username: $('#username-register').val(),
-        password: $('#password-register').val()
-    };
+    if (!$('#username-register').val() || !$('#password-register').val()) {
+        swal({
+            title: "Please Fill All Column!",
+            icon: "error",
+            buttons: false,
+            timer: 2000
+        });
+    } else {
+        const data = {
+            username: $('#username-register').val(),
+            password: $('#password-register').val()
+        };
 
-    $.ajax({
-        url: `http://localhost:3000/api/register`,
-        type: 'POST',
-        data
-
-    })
-        .done(function (token) {
-            console.log(token)
-            localStorage.setItem('token',token)
-            $('.login-form').hide()
-            $('.g-signin2').hide()
-            $('.register-form').hide()
-            $('#logout').show()
+        $.ajax({
+            url: `http://localhost:3000/api/register`,
+            type: 'POST',
+            data
         })
-        .fail(function (gg, textStatus) {
-            console.log('Err:', textStatus)
-        })
+            .done(function ({ token, username }) {
+                console.log(token, username)
+                localStorage.setItem('token', token)
+                localStorage.setItem('username', username)
+                hasLogin()
+                $('#username-register').val("")
+                $('#password-register').val("")
+                swal("Register Success!", `Welcome to the Todos App, ${username}`, "success");
+            })
+            .fail(function (gg, textStatus) {
+                console.log('Err:', textStatus)
+                swal({
+                    title: "Username already exist. Please register with another username!",
+                    icon: "error",
+                });
+            })
+    }
 }
 
-function back(){
+function back() {
     $("#logout").hide()
     $('.register-form').hide()
     $('.login-form').show()
     $('.g-signin2').show()
     $('.container').hide()
 }
-
-
-
